+++ service/src/no/mnemonic/act/platform/service/validators/Validator.java
@@ -5,6 +5,10 @@ package no.mnemonic.act.platform.service.validators;
  */
 public interface Validator {
 
+  enum ApplicableType {
+    FactType, ObjectType
+  }
+
   /**
    * Verifies that a value fulfills a specific format.
    *
@@ -12,5 +16,16 @@ public interface Validator {
    * @return True, if the value passes the validation.
    */
   boolean validate(String value);
+
+  /**
+   * Specifies whether the validator can be used for ObjectTypes, FactTypes or both.
+   * <p>
+   * The default implementation allows both. Implement this method if the usage should be restricted to only one type.
+   *
+   * @return Types which can use the validator
+   */
+  default ApplicableType[] appliesTo() {
+    return ApplicableType.values();
+  }
 
 }