+++ service/testsrc/no/mnemonic/act/platform/service/ti/handlers/ValidatorHandlerTest.java
@@ -1,6 +1,7 @@
 package no.mnemonic.act.platform.service.ti.handlers;
 
 import no.mnemonic.act.platform.api.exceptions.InvalidArgumentException;
+import no.mnemonic.act.platform.service.validators.Validator;
 import no.mnemonic.act.platform.service.validators.ValidatorFactory;
 import no.mnemonic.commons.utilities.collections.SetUtils;
 import org.junit.Before;
@@ -28,7 +29,8 @@ public class ValidatorHandlerTest {
 
   @Test
   public void testValidatorExists() throws InvalidArgumentException {
-    validatorHandler.assertValidatorExists("someValidator", "someParameter");
+    when(validatorFactory.get("someValidator", "someParameter")).thenReturn(createValidator(Validator.ApplicableType.FactType));
+    validatorHandler.assertValidator("someValidator", "someParameter", Validator.ApplicableType.FactType);
     verify(validatorFactory).get("someValidator", "someParameter");
   }
 
@@ -37,11 +39,57 @@ public class ValidatorHandlerTest {
     when(validatorFactory.get("someValidator", "someParameter")).thenThrow(IllegalArgumentException.class);
 
     InvalidArgumentException ex = assertThrows(
-      InvalidArgumentException.class,
-      () -> validatorHandler.assertValidatorExists("someValidator", "someParameter")
+            InvalidArgumentException.class,
+            () -> validatorHandler.assertValidator("someValidator", "someParameter", Validator.ApplicableType.FactType)
     );
     assertEquals(
-      SetUtils.set("validator.not.exist"),
-      SetUtils.set(ex.getValidationErrors(), InvalidArgumentException.ValidationError::getMessageTemplate));
+            SetUtils.set("validator.not.exist"),
+            SetUtils.set(ex.getValidationErrors(), InvalidArgumentException.ValidationError::getMessageTemplate));
+  }
+
+  @Test
+  public void testValidatorNotApplicable() {
+    when(validatorFactory.get("someValidator", "someParameter")).thenReturn(createValidator(Validator.ApplicableType.FactType));
+
+    InvalidArgumentException ex = assertThrows(
+            InvalidArgumentException.class,
+            () -> validatorHandler.assertValidator("someValidator", "someParameter", Validator.ApplicableType.ObjectType)
+    );
+    assertEquals(
+            SetUtils.set("validator.not.applicable"),
+            SetUtils.set(ex.getValidationErrors(), InvalidArgumentException.ValidationError::getMessageTemplate));
+  }
+
+  @Test
+  public void testValidatorApplicableFactType() throws InvalidArgumentException {
+    when(validatorFactory.get("someValidator", "someParameter")).thenReturn(createValidator(Validator.ApplicableType.FactType));
+    validatorHandler.assertValidator("someValidator", "someParameter", Validator.ApplicableType.FactType);
+  }
+
+  @Test
+  public void testValidatorApplicableObjectType() throws InvalidArgumentException {
+    when(validatorFactory.get("someValidator", "someParameter")).thenReturn(createValidator(Validator.ApplicableType.ObjectType));
+    validatorHandler.assertValidator("someValidator", "someParameter", Validator.ApplicableType.ObjectType);
+  }
+
+  @Test
+  public void testValidatorApplicableBoth() throws InvalidArgumentException {
+    when(validatorFactory.get("someValidator", "someParameter")).thenReturn(createValidator(Validator.ApplicableType.values()));
+    validatorHandler.assertValidator("someValidator", "someParameter", Validator.ApplicableType.FactType);
+    validatorHandler.assertValidator("someValidator", "someParameter", Validator.ApplicableType.ObjectType);
+  }
+
+  private Validator createValidator(Validator.ApplicableType... applicableType) {
+    return new Validator() {
+      @Override
+      public boolean validate(String value) {
+        return true;
+      }
+
+      @Override
+      public ApplicableType[] appliesTo() {
+        return applicableType;
+      }
+    };
   }
 }